simple-api/requirements.txt
firebase_admin
flask
google-cloud-pubsub
